agrobot/requirements.txt
openai==1.31.0
tiktoken==0.7.0
httpx==0.27.0
numpy==1.24.3
opencv-python==4.10.0.82
aiofiles==23.2.1
moviepy==1.0.3
pydub==0.25.1
aiohttp==3.9.5
requests==2.32.3
matplotlib==3.9.0
beautifulsoup4==4.12.3
pillow==10.3.0
pytube==15.0.0
pandas==2.2.2
seaborn==0.13.2
pymupdf4llm==0.0.5
openpyxl==3.1.3
aiogram==3.7.0
pymongo==4.7.3
motor==3.4.0
pydantic==2.7.3
fastapi==0.111.0
python-docx
python-pptx